refactor(cmd): 重构命令行标志相关代码
- 将 FloatFlag 重命名为 Float64Flag，以更准确地反映其类型
- 更新了相关的函数名称和文档，以与新的标志类型保持一致
- 删除了旧的 cmd_flags.go 文件，将其内容拆分到新的文件中
- 新增了专门针对不同数据类型的标志处理文件，提高了代码的可维护性

+++ cmd/cmd_flags_test.go
@@ -156,14 +156,14 @@ func TestFloatVar(t *testing.T) {
 				t.Error("FloatVar with nil pointer should panic")
 			}
 		}()
-		cmd.FloatVar(nil, "float", "f", 3.14, "test float flag")
+		cmd.Float64Var(nil, "float", "f", 3.14, "test float flag")
 	})
 
 	// 测试正常功能
 	t.Run("normal case", func(t *testing.T) {
 		cmd := NewCommand("test", "t", flag.ContinueOnError)
-		var floatFlag flags.FloatFlag
-		cmd.FloatVar(&floatFlag, "float", "fl", 3.14, "test float flag")
+		var floatFlag flags.Float64Flag
+		cmd.Float64Var(&floatFlag, "float", "fl", 3.14, "test float flag")
 
 		// 测试默认值
 		if floatFlag.Get() != 3.14 {
@@ -180,8 +180,8 @@ func TestFloatVar(t *testing.T) {
 
 		// 测试短标志解析
 		cmd = NewCommand("test", "t", flag.ContinueOnError)
-		var floatFlagShort flags.FloatFlag
-		cmd.FloatVar(&floatFlagShort, "float-short", "fs", 3.14, "test float short flag")
+		var floatFlagShort flags.Float64Flag
+		cmd.Float64Var(&floatFlagShort, "float-short", "fs", 3.14, "test float short flag")
 		if err := cmd.Parse([]string{"-fs", "1.618"}); err != nil {
 			t.Fatalf("Parse failed: %v", err)
 		}
